task: gh img style, mid: responsive nav
